fix: DB id issue
FCC wanted data set up so the object id generated by mongodb when a user was created is used as an identifier on exercises as well. This causes errors since each document needs to have a unique key as an _id. I used the UUID package to generate an id string.